deprecating hwinfo filesystem utils
